+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Perform unsolicited message check after putting activity into a queue.
 - Save content of remote emoji reactions to database.
 - Don't reject `Emoji` objects without `updated` property.
+- Don't store chain ID if subscription has remote recipient.
 
 ### Removed